move byline to guui
