+++ packages/guui/components/Byline/README.md
@@ -1,0 +1,7 @@
+Byline
+======
+
+### Import
+```js
+import { Byline } from '@guardian/guui/Byline/Byline';
+```
