+++ packages/guui/index.ts
@@ -4,3 +4,4 @@ export { Dropdown } from './components/Dropdown/Dropdown';
 export { BigNumber } from './components/BigNumber/BigNumber';
 export { Container } from './components/Container/Container';
 export { Cols as GridCols, Row as GridRow } from './components/Grid/Grid';
+export { Byline } from './components/Byline/Byline';
